Move lógica de view para views.py e melhora wagtailhooks.py
